rename test_crypto to vectors to avoid confusion
