+++ test/unit/internal/crypto/test_elliptic_curve.py
@@ -29,7 +29,7 @@ from aws_encryption_sdk.internal.crypto.elliptic_curve import (
     _ECCCurveParameters,
     generate_ecc_signing_key,
 )
-from .test_crypto import VALUES
+from .vectors import VALUES
 
 pytestmark = [pytest.mark.unit, pytest.mark.local]
